pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "preflight-ai"
version = "0.1.0"
description = "Intelligent pre-flight layer between developers and AI coding assistants"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Preflight AI"},
]
keywords = ["ai", "cli", "claude", "gemini", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "anthropic>=0.18",
    "google-genai>=1.0",
    "tiktoken>=0.5",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
preflight = "main:app"

[tool.setuptools]
py-modules = ["main"]
packages = ["core", "ai", "memory", "builders", "modes"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
